Getting data from others classes: Inside “onCreate” method: 1.	Creating a Bundle, in which I will pull the data (“key”, “value”) from others classes 2.	Checking if this bundle exists and it is not empty 3.	Making my integer from this class equals to the parameter from another class (using “key”)
Inside “onMapReady” method:
Checking
4.	 if my integer equals 1 => user choose capen library
5.	if my integer equals 1 => user choose lockwood library
6.	if my integer equals 1 => user choose music library

Created “ServiceMap” class, which will open a map on a background. I need it to receive (the distance and the duration)
Created “onActivityResult” method: I need it to receive the result (distance and duration)
Getting data from others classes:
Inside “onCreate” method:
1.	Creating a Bundle, in which I will pull the data (“key”, “value”) from others classes
2.	Checking if this bundle exists and it is not empty
3.	Making my integer from this class equals to the parameter from another class (using “key”)

Inside “onMapReady” method:
Checking
4.	 if my integer equals 1 => user choose capen library
5.	if my integer equals 1 => user choose lockwood library
6.	if my integer equals 1 => user choose music library

Created “ServiceMap” class, which will open a map on a background. I need it to receive the distance and the duration.
Created “onActivityResult” method: I need it to get the results and to sent the data to others classes.